Delete product by admin only

+++ src/controllers/productController.js
@@ -63,7 +63,16 @@ const updateProduct = async (req, res)=> {
 
 const deleteProduct = async (req, res) => {
     const id = req.params.id;
+    const user = req.user;
     try {
+        const product = await productService.getProductById(id);
+        
+        if(!product) return res.status(404).send("Product not found!");
+        
+        if(product.createdBy != user.id && !user.roles.includes(ROLE_ADMIN)) {
+            return res.status(403).send("Access denied!");
+        }
+
         await productService.deleteProduct(id);
         res.send(`Product deleted successfully of id: ${id}`)
     } catch (error) {
